+++ cpp.hint
@@ -9,3 +9,9 @@
 #define AddScriptFunctionNames(PARENT, __VA_ARGS__) std::vector<std::string> GetSupportedScriptFunctionNames() const override { std::vector<std::string> functionNames = PARENT::GetSupportedScriptFunctionNames(); functionNames.insert(functionNames.end(), {__VA_ARGS__}); return functionNames; }
 #define SerializableOverrideMethods int ReadProperty(const std::string_view &propName, Reader &reader) override; int Save(Writer &writer) const override;
 #define SerializableClassNameGetter const std::string & GetClassName() const override { return c_ClassName; }
+#define LuaBindingRegisterFunctionDeclarationForType(TYPENAME) static luabind::scope Register##TYPENAME##LuaBindings();
+#define LuaBindingRegisterFunctionDefinitionForType(OWNINGSCOPENAME, TYPENAME) luabind::scope OWNINGSCOPENAME##::Register##TYPENAME##LuaBindings()
+#define AbstractTypeLuaClassDefinition(TYPE, PARENTTYPE) luabind::class_<TYPE, PARENTTYPE>(#TYPE) .property("ClassName", &TYPE::GetClassName)
+#define ConcreteTypeLuaClassDefinition(TYPE, PARENTTYPE) luabind::class_<TYPE, PARENTTYPE>(#TYPE) .def("Clone", &Clone##TYPE, luabind::adopt(luabind::result)) .property("ClassName", &TYPE::GetClassName)
+#define RegisterLuaBindingsOfAbstractType(OWNINGSCOPENAME, TYPE) luabind::def((std::string("To") + std::string(#TYPE)).c_str(), (TYPE *(*)(Entity *))&To##TYPE), luabind::def((std::string("To") + std::string(#TYPE)).c_str(), (const TYPE *(*)(const Entity *))&ToConst##TYPE), OWNINGSCOPENAME##::Register##TYPE##LuaBindings()
+#define RegisterLuaBindingsOfConcreteType(OWNINGSCOPENAME, TYPE) luabind::def((std::string("Create") + std::string(#TYPE)).c_str(), (TYPE *(*)(std::string, std::string))&Create##TYPE, luabind::adopt(luabind::result)), luabind::def((std::string("Create") + std::string(#TYPE)).c_str(), (TYPE *(*)(std::string))&Create##TYPE, luabind::adopt(luabind::result)), luabind::def((std::string("Random") + std::string(#TYPE)).c_str(), (TYPE *(*)(std::string, int))&Random##TYPE, luabind::adopt(luabind::result)), luabind::def((std::string("Random") + std::string(#TYPE)).c_str(), (TYPE *(*)(std::string, std::string))&Random##TYPE, luabind::adopt(luabind::result)), luabind::def((std::string("Random") + std::string(#TYPE)).c_str(), (TYPE *(*)(std::string))&Random##TYPE, luabind::adopt(luabind::result)), luabind::def((std::string("To") + std::string(#TYPE)).c_str(), (TYPE *(*)(Entity *))&To##TYPE), luabind::def((std::string("To") + std::string(#TYPE)).c_str(), (const TYPE *(*)(const Entity *))&ToConst##TYPE), luabind::def((std::string("Is") + std::string(#TYPE)).c_str(), (bool(*)(const Entity *))&Is##TYPE), OWNINGSCOPENAME##::Register##TYPE##LuaBindings()
